% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unwrap_seminr_model}
\alias{.unwrap_seminr_model}
\title{Unwrap seminr Bootstrap Objects}
\usage{
.unwrap_seminr_model(model)
}
\arguments{
\item{model}{A seminr model (possibly bootstrapped).}
}
\value{
The unwrapped base model.
}
\description{
Extracts the base model from a bootstrapped seminr object.
}
\keyword{internal}
