% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssm_print.R
\name{print.fssm_assumptions}
\alias{print.fssm_assumptions}
\title{Print FSSM Assumptions}
\usage{
\method{print}{fssm_assumptions}(x, ...)
}
\arguments{
\item{x}{An fssm_assumptions object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print FSSM Assumptions
}
