% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_0_100}
\alias{rescale_0_100}
\title{Rescale Vector to 0-100 Range}
\usage{
rescale_0_100(x, min_val, max_val)
}
\arguments{
\item{x}{Numeric vector to rescale.}

\item{min_val}{Theoretical minimum value.}

\item{max_val}{Theoretical maximum value.}
}
\value{
Numeric vector rescaled to 0-100.
}
\description{
Simple linear rescaling using provided min/max values.
}
\keyword{internal}
